+++ samples/Components/TicTacComponent/App.fs
@@ -206,7 +206,7 @@ module App =
             let! model = Context.Mvu(program)
 
             Application() {
-                Window() {
+                Window(
                     ContentPage(
                         Grid(coldefs = [ Star ], rowdefs = [ Star; Auto; Auto ]) {
                             (Grid(coldefs = [ Star; Absolute 5.0; Star; Absolute 5.0; Star ], rowdefs = [ Star; Absolute 5.0; Star; Absolute 5.0; Star ]) {
@@ -264,6 +264,6 @@ module App =
                                 .gridRow(2)
                         }
                     )
-                }
+                )
             }
         }