+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 _No unreleased changes_
 
+## [9.0.0-pre5] - 2024-11-25
+
+### Changed
+- Use Fabulous 3.0.0-pre16
+
+### Removed
+- Remove all SingleChildBuilder signatures
+
 ## [9.0.0-pre4] - 2024-11-21
 
 ### Changed
@@ -291,7 +299,8 @@ Essentially v2.8.1 and v8.0.0 are similar except for the required .NET version.
 ### Changed
 - Fabulous.MauiControls has moved from the Fabulous repository to its own repository: [https://github.com/fabulous-dev/Fabulous.MauiControls](https://github.com/fabulous-dev/Fabulous.MauiControls)
 
-[unreleased]: https://github.com/fabulous-dev/Fabulous.MauiControls/compare/9.0.0-pre4...HEAD
+[unreleased]: https://github.com/fabulous-dev/Fabulous.MauiControls/compare/9.0.0-pre5...HEAD
+[9.0.0-pre5]: https://github.com/fabulous-dev/Fabulous.MauiControls/releases/tag/9.0.0-pre5
 [9.0.0-pre4]: https://github.com/fabulous-dev/Fabulous.MauiControls/releases/tag/9.0.0-pre4
 [9.0.0-pre3]: https://github.com/fabulous-dev/Fabulous.MauiControls/releases/tag/9.0.0-pre3
 [9.0.0-pre2]: https://github.com/fabulous-dev/Fabulous.MauiControls/releases/tag/9.0.0-pre2